Rough inital version of LBP predections [https://trello.com/c/NMWCQq5x]

+++ src/main/feature/pre-processing.h
@@ -2,6 +2,6 @@
 #define PreProcessing_H
 
 void illuminationInvariance(cv::Mat &image, cv::Mat &returnImage);
-void checkFaultLBP(int (&normalSampe)[5], int (&anomolySample)[5], cv::Mat &image);
+void checkFaultLBP(float (&normalSampe)[5], float (&anomolySample)[5], cv::Mat &image);
 
 #endif
