+++ src/main/feature/utils/beta/texture-beta.h
@@ -1,7 +1,7 @@
 #ifndef TextureBeta_H
 #define TextureBeta_H
 
-void computeLBP(cv::Mat &image, cv::Mat &LBPValues, int (&LBPHistogram)[5]);
+void computeLBP(cv::Mat &image, cv::Mat &LBPValues, float (&LBPHistogram)[5]);
 cv::Mat brigthen_darker_areas(const cv::Mat& img, int threshold, int amount);
 
 #endif
